pyproject.toml
[tool.poetry]
name = "django-auth"
version = "0.1.0"
description = ""
authors = ["Ilkka Kuisma <ikuisma@umich.edu>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.4"
django-crispy-forms = "^1.9.0"
django-webpack-loader = "^0.7.0"
djangorestframework = "^3.11.0"
social-auth-app-django = "^3.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
